Added a safe division method for BigDecimals (and implemented it)
